Change to immutable by default
